Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -2,7 +2,7 @@
 int main(){
     int a;
     scanf("%d",&a);
-    if (a>=2147483648){
+    if ((unsigned int)a>=2147483648){
         printf("Set");
     }
     else{
